+++ src/board.h
@@ -1,11 +1,12 @@
-//
-// Created by jakub on 28.07.2025.
+// s
+//  Created by jakub on 28.07.2025.
 //
 
 #pragma once
 #include <string>
 #include "move.h"
-class Board {
+class Board
+{
 private:
     char board[8][8] = {};
     char activeColor;
@@ -20,7 +21,12 @@ private:
 public:
     void startBoard();
     void printBoard() const;
-    bool isMoveValid(const Move& move);
-    void makeMove(const Move& move);
+    bool isMoveValid(const Move &move);
+    void makeMove(const Move &move);
+    std::string checkCastlingReason(const Move &move) const; // To dodane
 };
 
+//
+// Created by  chatGPT Macieja Soroki
+//
+// board.h
